Shopclues script fixed
