+++ data.json
@@ -1,12 +1,10 @@
 [
     {
-        "ORIGINAL_PRICE": "Rs.20,000", 
-        "NAME": "HTC Desire 828 Dual Sim (Pearl White, 16 GB)", 
-        "DISCOUNT": "43% off", 
-        "DISCOUNTED_PRICE": "Rs.11,999", 
-        "SALE_PRICE": "Rs.20,000", 
-        "OFFER": "Offer                            \r\n                                \r\n                                10% Cashback. Max Rs 200 on Vodafone m-pesa                                  T&C\r\n                                                                \r\n                                                        \r\n                                View offers\r\n                            \r\n                                                    \r\n                                                \r\n                                                    10% Cashback. Max Rs 200 on Vodafone m-pesa                                                                 T&C", 
-        "URL": "http://www.shopclues.com/htc-desire-828-dual-sim-pearl-white-16-gb-2.html", 
-        "PRODUCTID": "Product Id : 109923692"
+        "CATEGORY": "Computers & Accessories > Components > Memory", 
+        "ORIGINAL_PRICE": "2,404.00", 
+        "NAME": "Crucial CT4G4SFS8213 4GB 2133MHz DDR4 260-Pin Laptop Memory", 
+        "URL": "http://www.amazon.in/dp/B015HQ9SI6", 
+        "SALE_PRICE": "2,404.00", 
+        "AVAILABILITY": "In stock."
     }
 ]
