+++ datpaa.json
@@ -1,0 +1,9 @@
+[
+    {
+        "CATEGORY": null, 
+        "SALE_PRICE": "Rs.31,600", 
+        "URL": "http://www.shopclues.com/lg-nexus-5-16gb-18.html", 
+        "NAME": null, 
+        "ORIGINAL_PRICE": null
+    }
+]
