+++ mamamama.json
@@ -1,0 +1,12 @@
+[
+    {
+        "ORIGINAL_PRICE": null, 
+        "NAME": null, 
+        "DISCOUNT": null, 
+        "DISCOUNTED_PRICE": "Rs.31,600", 
+        "SALE_PRICE": null, 
+        "OFFER": "3 offers                            \r\n                                \r\n                                10% Instant Discount* on IndusInd Bank Credit Card                                  T&C\r\n                                                                \r\n                                                        \r\n                                View offers\r\n                            \r\n                                                    \r\n                                                \r\n                                                    10% Instant Discount* on IndusInd Bank Credit Card                                                                 T&C \r\n                                                            \r\n                                                    10% Cashback on Vodafone m-pesa, Max Cashback Rs 200                                                                 T&C \r\n                                                            \r\n                                                    Upto 100% Cashback On Mobikwik, Max Cashback Rs 500                                                                 T&C", 
+        "URL": "http://www.shopclues.com/lg-nexus-5-16gb-18.html", 
+        "PRODUCTID": "Product Id : 3158752"
+    }
+]
